+++ src/main/java/com/management/tab/infrastructure/jwt/JwtEncoder.java
@@ -1,6 +1,6 @@
-package com.management.tab.config.auth.security.jwt;
+package com.management.tab.infrastructure.jwt;
 
-import com.management.tab.config.auth.security.TokenEncoder;
+import com.management.tab.domain.auth.TokenEncoder;
 import com.management.tab.config.auth.security.config.properties.TokenProperties;
 import com.management.tab.config.auth.security.enums.TokenType;
 import com.nimbusds.jose.EncryptionMethod;
